+++ CHANGELOG.md
@@ -21,6 +21,9 @@ needed. This is equivalent to summarizing all activity on a feature branch versu
 
 ## Unreleased
 
+### Added
+* csv_2_supporting - Utility script which enables batch creation of supporting material files from a csv file with similar structure to the reagent_resources.csv plus two additional columns "Publications" and "Notes".
+
 ## v0.4.1
 
 ### Fixed